+++ src/View/AdminRoom/index.html.twig
@@ -12,7 +12,7 @@
 
                 <a role="button" class="btn btn-outline-primary btn-lg" href="/Admin/index">Retour</a>
 
-                <a role="button" class="btn btn-outline-primary btn-lg" href="/Admin/addAddress">Ajouter une adresse</a>
+                <a role="button" class="btn btn-outline-primary btn-lg" href="/AdminAddress/addAddress">Ajouter une adresse</a>
 
             </div>
